Put function in filter component
Doesnt totaly make sense but we dont have the grid component which it will probably work with. will interface a little bit with the filter compoent so might work here. Is there a place for stand alone functions. I thought utils might work for that but that raised an error